feature_engineering/cloud_functions/dataset_generator_http/requirements.txt
google-cloud-storage
google-cloud-datastore 
scikit-learn==0.20.3
opencv-python-headless==4.1.1.26
opencv-contrib-python==4.1.1.26
click==7.0
numpy==1.16.2
pandas==0.24.2
scikit-image==0.15.0
memory-profiler==0.55.0